+++ app/src/packages/work/gift/index.config.ts
@@ -1,0 +1,4 @@
+export default {
+  navigationStyle: 'custom',
+  navigationBarTitleText: '送你一卦',
+};